Fixing types in AppDelegate.

+++ TextDHxls/TextDHxls/AppDelegate.m
@@ -31,7 +31,7 @@
 	//[dhWS height:100 row:1];
 	[dhWS setWidth:10000 forColumn:0 defaultFormat:NULL];
 	
-	for(uint16_t idx=0; idx<10; ++idx) {
+	for(uint32_t idx=0; idx<10; ++idx) {
 		cell = [dhWS setCellAtRow:idx column:0 toString:[NSString stringWithFormat:@"Row %d", idx + 1]];
 		if(idx & 1) {
 			// prove we can get the cell reference later
@@ -45,18 +45,18 @@
 //	NSData *now = [NSDate date];
 //	NSDate *then = [NSDate dateWithString:@"1899-01-01 12:00:00 +0000").
 	
-	for(uint16_t idx=0; idx<10; ++idx) {
+	for(uint32_t idx=0; idx<10; ++idx) {
 		[dhWS setCellAtRow:idx column:1 toDoubleValue:3.1415f withNumberFormat:FMT_GENERAL + idx];
 	}
 
 	[dhWS setWidth:30000 forColumn:2 defaultFormat:NULL];
-	for(uint16_t idx=0; idx<7; ++idx) {
+	for(uint32_t idx=0; idx<7; ++idx) {
 		cell = [dhWS setCellAtRow:idx column:2 toString:@"Hello World"];
 		[cell setHorizontalAlignment:HALIGN_GENERAL + idx];
 	}
 
 	[dhWS setWidth:0xFFFF forColumn:3 defaultFormat:NULL];
-	for(uint16_t idx=0; idx<4; ++idx) {
+	for(uint32_t idx=0; idx<4; ++idx) {
 		[dhWS setHeight:24 forRow:idx defaultFormat:NULL];
 		cell = [dhWS setCellAtRow:idx column:3 toString:@"Hello World"];
 		[cell setVerticalAlignment:VALIGN_TOP + idx];
